Refactor health-check file-requirements UI into smaller modules
Split the ~970-line FileRequirementsContent into focused units: pure
report/card helpers under utils/fileRequirements, feedback/report-copy
hooks under hooks/, and one component per requirement card plus the
listing row and detail view under components/file_requirement. Relocate
the file-requirement types from the shared types.ts to their owning
util modules, and consolidate the feedback modal into EntryActions so
each call site no longer manages its own modal state. Behaviour is
preserved throughout.
