+++ src/renderer/src/extensions/health_check/utils/fileRequirements/fileRequirementActions.ts
@@ -1,9 +1,4 @@
 import { knownGames } from "@/extensions/gamemode_management/selectors";
-import type {
-  IDownloadedFile,
-  IFileRequirementCandidate,
-  IInstalledFile,
-} from "@/extensions/health_check/types";
 import { shouldShowPremiumAd } from "@/extensions/nexus_integration/selectors";
 import { nexusGames } from "@/extensions/nexus_integration/util";
 import { convertGameIdReverse } from "@/extensions/nexus_integration/util/convertGameId";
@@ -13,6 +8,9 @@ import { activeProfile } from "@/extensions/profile_management/selectors";
 import { log } from "@/logging";
 import type { IExtensionApi } from "@/types/IExtensionContext";
 import { opn, renderModName, sanitizeCSSId } from "@/util/api";
+
+import type { IDownloadedFile, IInstalledFile } from "./installedFiles";
+import type { IFileRequirementCandidate } from "./mapRequirementsReport";
 
 // File-level requirement actions: web links, single-file download, reveal-in-loadout,
 // and enable / switch active version.
